MatchTemplate algorithm successfully integrated
The matchTemplate algorithm was successfully integrated onto the project, therefore, it was now fully operational. 
The scaleFactor which was used to boost the recognition process was now commented to avoid unexpected errors when matching the template, however, the recognition process was largely slowed down. 
The threshold value (i.e. similarity) was set to 15. 
May require further re-deployment.